requirements.txt
python-telegram-bot==13.7
telebot==0.0.4
yt-dlp==2023.3.4
spotipy==2.19.0
mysql-connector-python==8.0.26
shazamio==0.1.0.1
requests==2.26.0
python-dotenv==0.19.0
